unit test added, default bd changed from mysql to hsql
