+++ src/main/resources/marvin.ini
@@ -1,4 +1,4 @@
-jdbc.driverClassName=com.mysql.jdbc.Driver
-jdbc.url=jdbc:mysql://localhost:3306/engine_metadata?useSSL=false
-jdbc.username=zhang
-jdbc.password=Mysql123@
+jdbc.driverClassName=org.hsqldb.jdbcDriver
+jdbc.url=jdbc:hsqldb:mem:engine_metadata
+jdbc.username=test
+jdbc.password=
